config sever, database, security and swagger

+++ backend/modules/src/main/java/sk/ytr/modules/dto/request/StudentRequestDTO.java
@@ -6,6 +6,7 @@ import sk.ytr.modules.constant.GenderTypeEnum;
 import sk.ytr.modules.entity.MedicalCampaign;
 import sk.ytr.modules.entity.Student;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -37,10 +38,10 @@ public class StudentRequestDTO {
     private String identityNumber;
 
     /** Cân nặng (kg) */
-    private Double weight;
+    private BigDecimal weight;
 
     /** Chiều cao (cm) */
-    private Double height;
+    private BigDecimal height;
 
     /** Nội dung thông báo về gia đình */
     private String notifyFamily;
